+++ R/cache.R
@@ -208,21 +208,18 @@ setMethod(
   definition = function(object, functionName) {
     cur <- current(object)
     if (NROW(cur)) {
-      if (cur$eventTime <= end(object)) {
-        whichCached <- grep(".useCache", object@params)
-        useCacheVals <- lapply(whichCached, function(x) {
-          object@params[[x]]$.useCache
-        })
+      whichCached <- grep(".useCache", object@params)
+      useCacheVals <- lapply(whichCached, function(x) {
+        object@params[[x]]$.useCache
+      })
 
-        whCurrent <- match(cur$moduleName, names(object@params)[whichCached])
-        if (isTRUE(useCacheVals[[whCurrent]])) {
-          cat(crayon::blue("  Using cached copy of", cur$moduleName, "module\n"))
-        } else {
-          cat(crayon::blue("  Using cached copy of", cur$eventType, "event in",
-                            cur$moduleName, "module\n"))
-        }
+      whCurrent <- match(cur$moduleName, names(object@params)[whichCached])
+      if (isTRUE(useCacheVals[[whCurrent]])) {
+        cat(crayon::blue("  Using cached copy of", cur$moduleName, "module\n"))
+      } else {
+        cat(crayon::blue("  Using cached copy of", cur$eventType, "event in",
+                         cur$moduleName, "module\n"))
       }
-
     } else {
       .cacheMessage(NULL, functionName)
     }
@@ -251,7 +248,10 @@ setMethod(
   ".checkCacheRepo",
   signature = "list",
   definition = function(object, create) {
+
+    object <- .findSimList(object)
     whSimList <- unlist(lapply(object, is, "simList"))
+
     if (any(whSimList)) {
       cacheRepo <- object[whSimList][[1]]@paths$cachePath # just take the first simList, if there are >1
     } else {
@@ -291,7 +291,8 @@ setMethod(
   signature = "simList",
   definition = function(object, cacheRepo, ...) {
     tmpl <- list(...)
-    whSimList <- which(unlist(lapply(tmpl, is, "simList")))
+    tmpl <- .findSimList(tmpl)
+    whSimList <- which(unlist(lapply(tmpl, is, "simList")))[1] # only take first simList -- may be a problem
     origEnv <- tmpl[[whSimList[1]]]@.envir
     isListOfSimLists <- if (is.list(object)) {
       if (is(object[[1]], "simList")) TRUE else FALSE
@@ -304,7 +305,7 @@ setMethod(
       for (i in seq_along(object)) {
         # need to keep the list(...) slots ...
         # i.e., Caching of simLists is mostly about objects in .envir
-        object2[[i]] <- Copy(list(...)[[whSimList]], objects = FALSE)
+        object2[[i]] <- Copy(tmpl[[whSimList]], objects = FALSE)
         object2[[i]]@.envir <- object[[i]]@.envir
         object2[[i]]@completed <- object[[i]]@completed
         object2[[i]]@simtimes <- object[[i]]@simtimes
@@ -318,9 +319,9 @@ setMethod(
         list2env(mget(lsOrigEnv[keepFromOrig], envir = tmpl[[whSimList]]@.envir),
                  envir = object2[[i]]@.envir)        }
     } else {
-      # need to keep the list(...) slots ...
+      # need to keep the tmpl slots ...
       # i.e., Caching of simLists is mostly about objects in .envir
-      object2 <- Copy(list(...)[[whSimList]], objects = FALSE)
+      object2 <- Copy(tmpl[[whSimList]], objects = FALSE)
       object2@.envir <- object@.envir
       object2@completed <- object@completed
       if (NROW(current(object2)) == 0) {
@@ -456,3 +457,27 @@ setMethod(
   definition = function(object) {
     object.size(as.list(object@.envir, all.names = TRUE)) + object.size(object)
 })
+
+
+#' Find simList in a nested list
+#'
+#' THis is recursive, so it will find the all simLists even if they are deeply nested.
+#'
+#' @param x any object, used here only when it is a list with at least one
+#'        \code{simList} in it
+#' @rdname findSimList
+.findSimList <- function(x) {
+  if (is.list(x)) {
+    out <- lapply(x, .findSimList)
+    out <- unlist(out, recursive = TRUE, use.names = FALSE) # get rid of NULL
+    if (is.null(out)) out <- list(NULL)
+  } else {
+    out <- inherits(x, "simList")
+    if (out) {
+      out <- x
+    } else {
+      out <- NULL
+    }
+  }
+  return(out)
+}