+++ tests/testthat/test-timeunits.R
@@ -190,6 +190,8 @@ test_that("timeunits with child and parent modules work correctly", {
   xxx1 <- gsub(xxx, pattern = 'timeunit = "year"', replacement = "timeunit = NA") # nolint
   cat(xxx1, file = fileName, sep = "\n")
 
+
+  ## child6 is used for further module testing
   suppressMessages(newModule("child6", tmpdir, open = FALSE))
   fileName <- file.path(tmpdir, "child6/child6.R")
   xxx <- readLines(fileName)
@@ -216,5 +218,52 @@ test_that("timeunits with child and parent modules work correctly", {
   mySimOut <- spades(mySim)
   expect_true(mySimOut$dp == file.path(dirname(fileName), "data"))
   expect_true(mySimOut$cm == file.path(modName))
+
+
+  ######
+  theFile <- file.path(tmpdir, "test")
+  write.table(x = data.frame(1), file = theFile)
+
+  xxx <- readLines(fileName)
+  modName <- basename(dirname(fileName))
+  lineOfInterest1 <- tail(grep(xxx, pattern = "expectsInput"),1)
+  xxx1 <- c(xxx[seq(lineOfInterest1-1)], "  expectsInput(\"b\", \"character\", \"temp thing\"),",
+            xxx[seq(length(xxx)-lineOfInterest1)+lineOfInterest1])
+  cat(xxx1, file = fileName, sep = "\n")
+
+  lineOfInterest <- grep(xxx1, pattern = ".inputObjects <- ")
+  xxx1 <- c(xxx1[seq(lineOfInterest - 1)], "  .inputObjects <- function(sim, a = asPath(file.path(inputPath(sim), \"test\"))) {",
+            "  sim$b <- a",
+            xxx1[seq(length(xxx1)-lineOfInterest)+lineOfInterest])
+  cat(xxx1, file = fileName, sep = "\n")
+
+  cacheDir <- file.path(tmpdir, "cache")
+  try(clearCache(cacheDir), silent = TRUE)
+  expect_silent(expect_message(mySim <- simInit(modules = list(modName),
+                                  paths = list(modulePath = tmpdir, inputPath = tmpdir, cachePath = cacheDir),
+                                  params = list("child6"= list(.useCache = ".inputObjects"))), all = TRUE, "Using or creating cached")
+  )
+
+  # pulls cached value
+  expect_message(expect_output(mySim <- simInit(modules = list(modName),
+                                 paths = list(modulePath = tmpdir, inputPath = tmpdir, cachePath = cacheDir),
+                   params = list("child6"= list(.useCache = ".inputObjects"))), regexp = "  Using cached"),
+                 all = TRUE, "Using or creating cached")
+  expect_true(identical(mySim$b, asPath(theFile)))
+
+  # Change the file that is in the arguments to .inputObjects
+  write.table(x = data.frame(sample(1e6,1)), file = theFile)
+  # Cache should force a rerun -- i.e., not cached value
+  expect_silent(expect_message(mySim <- simInit(modules = list(modName),
+                                                paths = list(modulePath = tmpdir, inputPath = tmpdir, cachePath = cacheDir),
+                                                params = list("child6"= list(.useCache = ".inputObjects"))), all = TRUE, "Using or creating cached")
+  )
+
+  # pulls cached value
+  expect_message(expect_output(mySim <- simInit(modules = list(modName),
+                                                paths = list(modulePath = tmpdir, inputPath = tmpdir, cachePath = cacheDir),
+                                                params = list("child6"= list(.useCache = ".inputObjects"))), regexp = "  Using cached"),
+                 all = TRUE, "Using or creating cached")
+
 
 })